+++ udata/core/dataset/forms.py
@@ -1,6 +1,10 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
+from urlparse import urlparse
+
+from flask import current_app
+
 from udata.forms import ModelForm, fields, validators
 from udata.i18n import lazy_gettext as _
 
@@ -10,7 +14,7 @@ from udata.core.spatial.forms import SpatialCoverageField
 from .models import (
     Dataset, Resource, License, Checksum, CommunityResource,
     UPDATE_FREQUENCIES, DEFAULT_FREQUENCY, RESOURCE_FILETYPES, CHECKSUM_TYPES,
-    LEGACY_FREQUENCIES, RESOURCE_TYPES
+    LEGACY_FREQUENCIES, RESOURCE_TYPES, RESOURCE_FILETYPE_FILE,
 )
 
 __all__ = ('DatasetForm', 'ResourceForm', 'CommunityResourceForm')
@@ -29,11 +33,28 @@ def normalize_format(data):
         return data.strip().lower()
 
 
+def enforce_filetype_file(form, field):
+    '''Only allowed domains in resource.url when filetype is file'''
+    if field.data != RESOURCE_FILETYPE_FILE:
+        return
+    url = form._fields.get('url').data
+    domain = urlparse(url).netloc
+    allowed_domains = current_app.config.get(
+        'RESOURCES_FILE_ALLOWED_DOMAINS', [])
+    allowed_domains += [current_app.config.get('SERVER_NAME')]
+    if '*' in allowed_domains:
+        return
+    if domain and domain not in allowed_domains:
+        message = 'URL domain not allowed for filetype {}'.format(
+            RESOURCE_FILETYPE_FILE)
+        raise validators.ValidationError(_(message))
+
+
 class BaseResourceForm(ModelForm):
     title = fields.StringField(_('Title'), [validators.required()])
     description = fields.MarkdownField(_('Description'))
     filetype = fields.RadioField(
-        _('File type'), [validators.required()],
+        _('File type'), [validators.required(), enforce_filetype_file],
         choices=RESOURCE_FILETYPES.items(), default='file',
         description=_('Whether the resource is an uploaded file, '
                       'a remote file or an API'))