+++ udata/settings.py
@@ -214,6 +214,9 @@ class Defaults(object):
         # RDF
         'rdf', 'ttl', 'n3',
     ]
+    # Whitelist of urls domains for resource with filetype `file`
+    # SERVER_NAME is always included, `*` is a supported value (wildcard)
+    RESOURCES_FILE_ALLOWED_DOMAINS = tuple()
 
     # How much time upload chunks are kept before cleanup
     UPLOAD_MAX_RETENTION = 24 * HOUR
@@ -272,6 +275,7 @@ class Testing(object):
     LOGGER_HANDLER_POLICY = 'never'
     CELERYD_HIJACK_ROOT_LOGGER = False
     USE_METRICS = False
+    RESOURCES_FILE_ALLOWED_DOMAINS = ['*']
 
 
 class Debug(Defaults):